requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
Werkzeug==3.1.4

# --- CÁC MỤC QUAN TRỌNG CẦN BẬT/THÊM ---
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0